Include Grafana tests in npm test
