+++ tests/fast-integration/monitoring/.env.template
