+++ tests/fast-integration/test/fixtures/getting-started-guide/index.js
