+++ tests/fast-integration/test/fixtures/getting-started-guide/microservice.yaml
