+++ tests/fast-integration/test/fixtures/getting-started-guide/ns.yaml
