+++ tests/fast-integration/test/fixtures/getting-started-guide/redis-addon-sb-si.yaml
