+++ tests/fast-integration/test/fixtures/getting-started-guide/redis-sbu.yaml
