+++ tests/fast-integration/test/fixtures/getting-started-guide/xf-mocks.yaml
