+++ tests/fast-integration/test/index.js
@@ -1,4 +1,5 @@
 module.exports = {
     ...require('./1-commerce-mock'),
-    ...require('./2-getting-started-guides'),
+    ...require('./2-getting-started-guide'),
+    ...require('./3-observability'),
 }
